# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CTU Theseus API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/sweeps": {
      "post": {
        "tags": [
          "sweeps"
        ],
        "summary": "Create Sweep",
        "operationId": "create_sweep",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSweepBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SweepCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sweeps"
        ],
        "summary": "List Sweeps",
        "operationId": "list_sweeps",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SweepListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sweeps/{sweep_id}": {
      "get": {
        "tags": [
          "sweeps"
        ],
        "summary": "Get Sweep",
        "operationId": "get_sweep",
        "parameters": [
          {
            "name": "sweep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sweep Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SweepDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sweeps/{sweep_id}/cancel": {
      "post": {
        "tags": [
          "sweeps"
        ],
        "summary": "Cancel Sweep",
        "operationId": "cancel_sweep",
        "parameters": [
          {
            "name": "sweep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sweep Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSweepBody": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "datasetVersionId": {
            "type": "string",
            "format": "uuid",
            "title": "Datasetversionid"
          },
          "searchSpace": {
            "$ref": "#/components/schemas/SearchSpace"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "grid",
              "random"
            ],
            "title": "Strategy"
          },
          "maxTrials": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Maxtrials"
          }
        },
        "type": "object",
        "required": [
          "name",
          "datasetVersionId",
          "searchSpace",
          "strategy",
          "maxTrials"
        ],
        "title": "CreateSweepBody"
      },
...
      "EvaluationBrief": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "failed"
            ],
            "title": "Status"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "macroF1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macrof1"
          }
        },
        "type": "object",
        "required": [
          "status",
          "accuracy",
          "macroF1"
        ],
        "title": "EvaluationBrief",
        "description": "Denormalized accuracy / macro-F1, so lists and comparisons need no per-run round trip."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunRow": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled"
            ],
            "title": "Status"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "title": "Projectid"
          },
          "datasetVersionId": {
            "type": "string",
            "format": "uuid",
            "title": "Datasetversionid"
          },
          "sweepId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sweepid"
          },
          "trialIndex": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trialindex"
          },
          "hyperparameters": {
            "title": "Hyperparameters"
          },
          "bestEpoch": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bestepoch"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "failedMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failedmessage"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "projectId",
          "datasetVersionId",
          "sweepId",
          "trialIndex",
          "hyperparameters",
          "bestEpoch",
          "startedAt",
          "completedAt",
          "failedMessage",
          "createdAt",
          "updatedAt"
        ],
        "title": "RunRow"
      },
...
      "SearchSpace": {
        "properties": {
          "epochs": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Epochs"
          },
          "batchSize": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "string",
                      "const": "auto"
                    }
                  ]
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Batchsize"
          },
          "learningRate": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Learningrate"
          },
          "earlyStopPatience": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Earlystoppatience"
          },
          "encoderId": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Encoderid"
          }
        },
        "type": "object",
        "title": "SearchSpace",
        "description": "One candidate-value list per sweepable trainer knob (camelCase keys, as stored and expanded)."
      },
...
      "SweepCreatedResponse": {
        "properties": {
          "sweep": {
            "$ref": "#/components/schemas/SweepRow"
          },
          "trials": {
            "items": {
              "$ref": "#/components/schemas/RunRow"
            },
            "type": "array",
            "title": "Trials"
          }
        },
        "type": "object",
        "required": [
          "sweep",
          "trials"
        ],
        "title": "SweepCreatedResponse"
      },
      "SweepDetailResponse": {
        "properties": {
          "sweep": {
            "$ref": "#/components/schemas/SweepRow"
          },
          "trials": {
            "items": {
              "$ref": "#/components/schemas/SweepTrial"
            },
            "type": "array",
            "title": "Trials"
          }
        },
        "type": "object",
        "required": [
          "sweep",
          "trials"
        ],
        "title": "SweepDetailResponse"
      },
      "SweepListResponse": {
        "properties": {
          "sweeps": {
            "items": {
              "$ref": "#/components/schemas/SweepSummary"
            },
            "type": "array",
            "title": "Sweeps"
          }
        },
        "type": "object",
        "required": [
          "sweeps"
        ],
        "title": "SweepListResponse"
      },
      "SweepRow": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "title": "Projectid"
          },
          "datasetVersionId": {
            "type": "string",
            "format": "uuid",
            "title": "Datasetversionid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "searchSpace": {
            "title": "Searchspace"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "grid",
              "random"
            ],
            "title": "Strategy"
          },
          "maxTrials": {
            "type": "integer",
            "title": "Maxtrials"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "canceled"
            ],
            "title": "Status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "projectId",
          "datasetVersionId",
          "name",
          "searchSpace",
          "strategy",
          "maxTrials",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "SweepRow"
      },
      "SweepSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "grid",
              "random"
            ],
            "title": "Strategy"
          },
          "maxTrials": {
            "type": "integer",
            "title": "Maxtrials"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "canceled"
            ],
            "title": "Status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "trialCount": {
            "type": "integer",
            "title": "Trialcount"
          },
          "completedTrialCount": {
            "type": "integer",
            "title": "Completedtrialcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "strategy",
          "maxTrials",
          "status",
          "createdAt",
          "trialCount",
          "completedTrialCount"
        ],
        "title": "SweepSummary"
      },
      "SweepTrial": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled"
            ],
            "title": "Status"
          },
          "hyperparameters": {
            "title": "Hyperparameters"
          },
          "trialIndex": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trialindex"
          },
          "failedMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failedmessage"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "hyperparameters",
          "trialIndex",
          "failedMessage",
          "createdAt",
          "completedAt"
        ],
        "title": "SweepTrial"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}